+++ Runtime/FloatAnimatorParameter.cs
@@ -1,0 +1,11 @@
+namespace dGameBoy101b.AnimatorParameters
+{
+	public class FloatAnimatorParameter : AnimatorParameter<float>
+	{
+		public override float Value 
+		{
+			get => this.Animator.GetFloat(this.ParameterName);
+			set => this.Animator.SetFloat(this.ParameterName, value);
+		}
+	}
+}